svn98
intersect:
+fixed 0/0 case in bsp node traversal
+autodetect DELTA_INSIDE_PLANE, prepared but not turned on (should fix bunny)
+robustness tests in FASTEST (define TEST)
 investigating precision, any kdnode makes clear_miss increase from 1:30000 to 1:100

+++ src/BunnyTest/plymeshreader.cpp
@@ -33,7 +33,7 @@ static int vertex_callback(p_ply_argument argument)
 	long vertindex;
 	ply_get_argument_element(argument, NULL, &vertindex);
 
-	current_mesh->verts[vertindex].pos[current_cmpnt] = ply_get_argument_value(argument);
+	current_mesh->verts[vertindex].pos[current_cmpnt] = (float)ply_get_argument_value(argument);
 
 
     return 1;
